Refactor: Update package imports to align with `v1.stats` changes
