+++ agent/src/main/kotlin/dev/httpmarco/polocloud/agent/stats/StatsGrpcService.kt
@@ -4,9 +4,9 @@ import dev.httpmarco.polocloud.agent.Agent
 import dev.httpmarco.polocloud.common.os.cpuUsage
 import dev.httpmarco.polocloud.common.os.usedMemory
 import dev.httpmarco.polocloud.shared.stats.Stats
-import dev.httpmarco.polocloud.v1.groups.GetStatsRequest
-import dev.httpmarco.polocloud.v1.groups.GetStatsResponse
-import dev.httpmarco.polocloud.v1.groups.StatsControllerGrpc
+import dev.httpmarco.polocloud.v1.stats.GetStatsRequest
+import dev.httpmarco.polocloud.v1.stats.GetStatsResponse
+import dev.httpmarco.polocloud.v1.stats.StatsControllerGrpc
 import io.grpc.stub.StreamObserver
 
 class StatsGrpcService : StatsControllerGrpc.StatsControllerImplBase() {
